FTP Formatter
This file formats a text file full of FTP sites in the format of ftp://foo.ac.uk/ down to foo.ac.uk so that it can be scanned and then archived easily. Use it in this format 
FTPArchiveFormatter.py TextFileOfFTP.txt

In the same folder, after the script has done, you will have a file called Results.txt that will have all the servers formatted right so that you can import it right into the scanner and get the size of the FTP for downloading by the warriors.